+++ packages/storybook-html/stories/components/story-card/story-card.stories.js
@@ -1,10 +1,9 @@
 import "@uqds/card/src/scss/main.scss";
 import "@uqds/icon/src/scss/main.scss";
-import { cardDecorator } from "../../../lib/decorators";
+import { cardDecorator, storyCardDecorator } from "../../../lib/decorators";
 
 export default {
   title: "Components/Cards/Story Card",
-  decorators: [cardDecorator],
   argTypes: {
     title: "text",
     href: "text",
@@ -13,7 +12,7 @@ export default {
     bottomLabel: "text",
     variant: {
       control: "select",
-      options: ["story", "news", "event"],
+      options: ["story", "news", "event", "landscape", "feature"],
     },
     image: {
       control: "select",
@@ -23,7 +22,7 @@ export default {
   args: {
     title: "How does ATAR work and what does it mean?",
     description:
-      "The University of Queensland is honouring a learning scientist who is leading by example to inspire and teach future teachers.",
+      "The University of Queensland is honouring a learning scientist who is leading by example to inspire and teach future teachers. The University of Queensland is honouring a learning scientist who is leading by example to inspire and teach future teachers. The University of Queensland is honouring a learning scientist who is leading by example to inspire and teach future teachers.",
     topLabel: "Uni life",
     bottomLabel: "3 minute read time",
     image: "images/card/card-example-herston-campus.jpg",
@@ -36,6 +35,7 @@ export const StoryCard = {
     variant: "story",
     description: "",
   },
+  decorators: [cardDecorator],
   render: ({
     variant,
     image,
@@ -84,5 +84,55 @@ export const EventCard = {
     bottomLabel: "St Lucia",
     variant: "event",
     description: "",
+  },
+};
+
+export const LandscapeCard = {
+  ...StoryCard,
+  decorators: [storyCardDecorator],
+  args: {
+    title:
+      "ATAR advice for school leavers from Australia’s University Teacher of the Year ATAR advice for school leavers from Australia’s University Teacher of the Year",
+    topLabel: "News",
+    bottomLabel: "20 September 2023",
+    variant: "landscape",
+  },
+};
+
+export const LandscapeMissingImageCard = {
+  ...LandscapeCard,
+  args: {
+    topLabel: "News",
+    bottomLabel: "20 September 2023",
+    variant: "landscape",
+  },
+  render: ({ variant, title, href, topLabel, description, bottomLabel }) => `
+<div class="uq-story-card uq-story-card--${variant}">
+  <div class="uq-story-card__image"></div>
+  <div class="uq-story-card__content">
+    <div class="uq-story-card__top-label">${topLabel}</div>
+    <h3 class="uq-story-card__title"><a href="${href}">${title}</a></h3>
+    <div class="uq-story-card__description">${description}</div>
+    <div class="uq-story-card__bottom-label">${bottomLabel}</div>
+  </div>
+</div>`,
+};
+
+export const LandscapeNoImageCard = {
+  ...LandscapeCard,
+  args: {
+    topLabel: "News",
+    bottomLabel: "20 September 2023",
+    variant: "landscape",
+    image: null,
+  },
+};
+
+export const FeatureCard = {
+  ...LandscapeCard,
+  args: {
+    topLabel: "News",
+    bottomLabel: "20 September 2023",
+    variant: "feature",
   },
 };