+++ packages/body/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4-alpha.2](https://github.com/uq-its-ss/design-system/compare/@uqds/body@0.1.4-alpha.1...@uqds/body@0.1.4-alpha.2) (2024-12-16)
+
+### Bug Fixes
+
+- ITSADSSD-59530: Search bar QA ([#581](https://github.com/uq-its-ss/design-system/issues/581)) ([1625632](https://github.com/uq-its-ss/design-system/commit/16256322335dddba7766d4309bd09c6109c6a888))
+
+## [0.1.4-alpha.1](https://github.com/uq-its-ss/design-system/compare/@uqds/body@0.1.4-alpha.0...@uqds/body@0.1.4-alpha.1) (2024-12-05)
+
+**Note:** Version bump only for package @uqds/body
+
 ## [0.1.4-alpha.0](https://github.com/uq-its-ss/design-system/compare/@uqds/body@0.1.3...@uqds/body@0.1.4-alpha.0) (2024-09-04)
 
 **Note:** Version bump only for package @uqds/body