+++ packages/header/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4-alpha.4](https://github.com/uq-its-ss/design-system/compare/@uqds/header@1.2.4-alpha.3...@uqds/header@1.2.4-alpha.4) (2024-12-19)
+
+**Note:** Version bump only for package @uqds/header
+
+## [1.2.4-alpha.3](https://github.com/uq-its-ss/design-system/compare/@uqds/header@1.2.4-alpha.2...@uqds/header@1.2.4-alpha.3) (2024-12-16)
+
+**Note:** Version bump only for package @uqds/header
+
+## [1.2.4-alpha.2](https://github.com/uq-its-ss/design-system/compare/@uqds/header@1.2.4-alpha.1...@uqds/header@1.2.4-alpha.2) (2024-12-12)
+
+**Note:** Version bump only for package @uqds/header
+
+## [1.2.4-alpha.1](https://github.com/uq-its-ss/design-system/compare/@uqds/header@1.2.4-alpha.0...@uqds/header@1.2.4-alpha.1) (2024-12-05)
+
+**Note:** Version bump only for package @uqds/header
+
 ## [1.2.4-alpha.0](https://github.com/uq-its-ss/design-system/compare/@uqds/header@1.2.3...@uqds/header@1.2.4-alpha.0) (2024-09-04)
 
 **Note:** Version bump only for package @uqds/header