+++ packages/pane/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.0](https://github.com/uq-its-ss/design-system/compare/@uqds/pane@0.1.4-alpha.1...@uqds/pane@0.2.0-alpha.0) (2024-12-05)
+
+### Features
+
+- ITSADSSD-59302: Video pane ([#569](https://github.com/uq-its-ss/design-system/issues/569)) ([9ae841a](https://github.com/uq-its-ss/design-system/commit/9ae841a5987062a149fa12baa61e92c45b026a6c))
+- ITSADSSD-59777: Accordion updates and article template ([#567](https://github.com/uq-its-ss/design-system/issues/567)) ([8311fa4](https://github.com/uq-its-ss/design-system/commit/8311fa49bc079806e221931f6abd3ec948e0b9fe))
+
 ## [0.1.4-alpha.1](https://github.com/uq-its-ss/design-system/compare/@uqds/pane@0.1.4-alpha.0...@uqds/pane@0.1.4-alpha.1) (2024-09-24)
 
 **Note:** Version bump only for package @uqds/pane